Fetching route parameters reactively.
The thing about snapshot routing (this.route.snapshot.params['id'] for example) is that as name
implies it only saves a snapshot of url when a component is loaded (maaybe being inside ngOnInit
function has a liiiiitle to do with that ;). If there would be a navigation on that very same
component leading to it Angular will ignore it, thinking that there's no point reloading/refreshing a
component that is already loaded. So with a link inside user component pointing to another user then
current one, the url in browser will change but commponent won't give a fuck and stay the same.

To prevent that from happening you place a subcribe call to route.params (no snapshot part here).
That way each time url changes the function inside subscribe will be notified and "address" the issue
:>

Notice the import of Params next to ActivatedRoute in order for this to work.