About destroying route observables
In the last example all's fine and will work peachy. No more needs to be done. But not every
observable will be so kind, especially if you attach your own later down the road. Thing is that
subscription function tends to live away from component in memory. Component being destroyed does not
have to equal subscription being destroyed and that could get messy. In this example angular does
handle this automatically, but in other cases it won't. So what you need to do is to set up a
Destructor function to clean it up.

Here's that role is played by onDestroy function (imported from @angular/core) that should be
implemented with OnDestroy interface as well.
Additionally you'll be importing Subscription from rxjs. Now use the subscription to create private
property to grab the subscription function from ngOnInit, and then pass it to onDestroy() function to
clean it up once you're done with the component.