redirecting routes and wildcards.
Not all paths need to have components attached. Some can simply redirect to others that do. This is using redirectTo option instead
of "component" option in a given path, like so:

{path: 'not-found', component:PageNotFoundComponent},
{path: '**', redirectTo:'/not-found'}

Notice the "**" ? That's the wildcard match. It matches any route a user enters. That's why it's great for catching invalid urls,
that have no routes in our application. That is also the reason it MUST be LAST. It'll match any path, valid or invalid. That's why the
proper paths must be given chance to be read and used first. If you place it on top, every path, good or bad would be redirected to "not-found" path.